Rename getUserByLoginToken to getUserIdByLoginToken
This makes sure we only get the userId and no other details from the
user. That is something for developers to decide and more in line with
DDP itself. See reasoning in
https://github.com/Swydo/ddp-apollo/pull/188
