+++ lib/server/getUserIdByLoginToken.js
@@ -1,7 +1,7 @@
 const USER_TOKEN_PATH = 'services.resume.loginTokens.hashedToken';
 export const NO_VALID_USER_ERROR = new Error('NO_VALID_USER');
 
-export async function getUserByLoginToken(loginToken) {
+export async function getUserIdByLoginToken(loginToken) {
   // `accounts-base` is a weak dependency, so we'll try to require it
   // eslint-disable-next-line global-require, prefer-destructuring
   const { Accounts } = require('meteor/accounts-base');
@@ -31,5 +31,5 @@ export async function getUserByLoginToken(loginToken) {
 
   if (isTokenExpired) { throw NO_VALID_USER_ERROR; }
 
-  return user;
+  return user._id;
 }
