+++ lib/server/meteorAuthMiddleware.js
@@ -1,4 +1,4 @@
-import { getUserByLoginToken, NO_VALID_USER_ERROR } from './getUserByLoginToken';
+import { getUserIdByLoginToken, NO_VALID_USER_ERROR } from './getUserIdByLoginToken';
 
 export function meteorAuthMiddleware(req, res, next) {
   let tokenType;
@@ -16,9 +16,9 @@ export function meteorAuthMiddleware(req, res, next) {
   }
 
   // get the user for the context
-  getUserByLoginToken(loginToken)
-    .then((user) => {
-      req.userId = user && user._id;
+  getUserIdByLoginToken(loginToken)
+    .then((userId) => {
+      req.userId = userId;
       next();
     })
     .catch((err) => {
