+++ specs/server.spec.js
@@ -5,4 +5,4 @@ import './server/helpers/setupLoginByUserId';
 // specs
 import './server/server.spec';
 import './server/setup.spec';
-import './server/getUserByLoginToken.spec';
+import './server/getUserIdByLoginToken.spec';
